requirements.txt
langchain
langchain-openai
python-dotenv
langchain-community
langchain_core
streamlit